+++ src/eval-sandbox.js
@@ -4,17 +4,27 @@ const Sandbox = require('./lib/sandbox.js')
 const ArrayUtils = require('./lib/array-utils.js')
 
 class EvalSandbox {
-  constructor(parent, makeGlobal) {
+  constructor(parent, makeGlobal, userProps = []) {
     this.makeGlobal = makeGlobal
     this.sandbox = Sandbox(parent)
     this.parent = parent
     var properties = Object.keys(parent)
     properties.forEach((property) => this.add(property))
+    this.userProps = userProps
   }
 
   add(name) {
     if(this.makeGlobal) window[name] = this.parent[name]
     this.sandbox.addToContext(name, `parent.${name}`)
+  }
+
+  tick() {
+    if(this.makeGlobal) {
+      this.userProps.forEach((property) => {
+        this.parent[property] = window[property]
+      })
+      //  this.parent.speed = window.speed
+    }
   }
 
   eval(code) {