+++ src/eval-sandbox.js
@@ -18,6 +18,15 @@ class EvalSandbox {
     this.sandbox.addToContext(name, `parent.${name}`)
   }
 
+// sets on window as well as synth object if global (not needed for objects, which can be set directly)
+
+  set(property, value) {
+    if(this.makeGlobal) {
+      window[property] = value
+      parent[property] = value
+    }
+  }
+
   tick() {
     if(this.makeGlobal) {
       this.userProps.forEach((property) => {